requirements.txt

aiofiles>=24.1.0
aiohttp>=3.11.0
asyncio>=3.4.3
asyncssh>=2.14.2
dnspython>=2.7.0
email-validator>=2.2.0
flask>=3.1.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.1
gunicorn>=23.0.0
paramiko>=3.5.1
pillow>=10.2.0
psutil>=5.9.0
psycopg2-binary>=2.9.10
pydantic>=2.11.0
pymongo[srv]==4.6.1
python-dotenv>=1.1.0
pytz>=2024.1
requests>=2.31.0
sqlalchemy>=2.0.27
motor
werkzeug
wtforms
aiohttp>=3.7.4
asyncssh
motor
pymongo[srv]==4.6.1
python-dotenv
aiofiles
asyncio
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
paramiko
pillow
psutil
psycopg2-binary
pydantic
pytz
requests
sqlalchemy
psutil
aiohttp
asyncssh
python-dotenv
asyncssh
python-dotenv
aiohttp
asyncssh
email-validator
flask
flask-login
flask-sqlalchemy
gunicorn
motor
pillow
psutil
psycopg2-binary
pymongo[srv]
python-dotenv
aiofiles
asyncssh
motor
paramiko
pillow
psutil
pydantic
pymongo[srv]
python-dotenv
pytz
requests
discord-py
py-cord
motor
py-cord
pymongo
python-dotenv
discord.py
bson
py-cord
discord.py
dnspython
pymongo[srv]>=4.3
dnspython
motor
py-cord
pydantic
pymongo
python-dotenv
pytz
aiofiles>=24.1.0
aiohttp>=3.11.0
asyncio>=3.4.3
asyncssh>=2.14.2
dnspython>=2.7.0
motor>=3.3.2
paramiko>=3.5.1
pillow>=10.2.0
psutil>=5.9.0
py-cord==2.6.1
pydantic>=2.11.0
pymongo[srv]>=4.6.1
python-dotenv>=1.1.0
pytz>=2024.1
